fix: repair the defects a wider lint and a type checker surfaced
Turning on more rules was meant to be hygiene. Four of the findings were real.

**The knowledge graph was discarding every stored importance and confidence.**
On a `sqlite3.Row`, `"key" in row` tests the row's VALUES, not its keys — so
`row["importance"] if "importance" in row else 0.5` was almost always false, and
the default was used instead of the number that had been written. Measured:
importance stuck at 0.55 where four mentions should have reached 0.7, and a
0.97-confidence entity read back as 0.8. The columns are NOT NULL with defaults
in the schema and are backfilled by `_migrate`, so they are always present and
can simply be read. `tests/test_sqlite_row_membership.py` pins both the language
behaviour and the fix, and SIM118/SIM401 are disabled in pyproject because their
autofixes reintroduce exactly this.

**A sort key computed as `0 if x else 0`.** The canonical recall tiebreaker
never ordered anything. Removed rather than guessed at: whether time-sensitive
claims should win a tie is a product question, and dropping a constant from a
stable sort changes no output.

**`proc.stdin.write()` on a stdin that can be None**, when handing Claude an
authorization code. Now says the sign-in is not waiting for one.

**A regex searched twice, the second time unguarded** — `parse_date_range` found
`(past|last) N days` and then re-searched with a looser pattern before calling
`.group()` on the result. Reuses the match.

Also: connector capabilities are duck-typed, so `send_email`, `create_event` and
`search_and_ingest` are checked before they are called — otherwise a source
connected read-only produces an AttributeError and a 500 with a traceback in it,
which the product rules forbid surfacing. The live entity-type feed is
constrained to the allowlist where it is produced, not only inside
`upsert_entity`, because that list is rendered straight into the page.

**Dead code removed** (87 lines, each checked by hand for dynamic references and
for a documented contract before deleting): `Brain._is_prose`, superseded by
`_FULL_SOURCES`; `contained_secret`, for telemetry that never happened;
`first_build_from_messages`; five unused `CanonicalStore` methods;
`MemoryStore.delete_source`; `ProviderConnection.status_for`.

The rest is mechanical: `raise ... from`, lazy `%s` logging, iterator unpacking,
`zip(strict=True)` where a short batch would otherwise write vectors onto the
wrong rows, and annotations where `list` as a method name shadowed the builtin
and made every `-> list[X]` in three stores mean "returns the method".

+++ lodestone/hud.py
@@ -257,7 +257,7 @@ class _Bridge:
         if nswin is None:
             return False
         try:
-            wanted = max(_MIN_HEIGHT, min(_MAX_HEIGHT, int(round(float(height)))))
+            wanted = max(_MIN_HEIGHT, min(_MAX_HEIGHT, round(float(height))))
         except (TypeError, ValueError):
             return False
         from PyObjCTools import AppHelper
